refactor(ui): extract api/sync.ts — sync, peer, and actor endpoints
Move the 16 /api/sync/* viewer endpoints into api/sync.ts: status,
invite import, peer lifecycle (accept, rename, delete, scope,
identity, assign actor), actor CRUD (create, rename, merge,
deactivate), legacy-device claim, and the manual sync-now trigger.
The barrel re-exports each function so existing `api.loadSyncStatus`
etc. call sites stay unchanged.
